fix: 크루 리포트 언어 사용 비율 반환 시 double에서 int 타입으로 변경

+++ src/main/java/com/gsm/blabla/crew/dto/CrewReportResponseDto.java
@@ -18,11 +18,11 @@ public class CrewReportResponseDto {
     private List<MemberResponseDto> members;
     private String bubbleChart;
     private List<Map<String, Object>> keyword;
-    private Map<String, Double> languageRatio;
+    private Map<String, Integer> languageRatio;
     private List<MemberResponseDto> feedbacks;
 
     public static CrewReportResponseDto crewReportResponse(Map<String, String> info, List<MemberResponseDto> members,
-        String bubbleChart, List<Map<String, Object>> keyword, Map<String, Double> languageRatio,
+        String bubbleChart, List<Map<String, Object>> keyword, Map<String, Integer> languageRatio,
         List<MemberResponseDto> feedbacks) {
         return CrewReportResponseDto.builder()
             .info(info)
